fix: apply council fixes — extract config-utils, normalize Tauri file open
- Extract stateToEngineConfig/engineConfigToState to shared src/core/config-utils.js
- Add isFileFormatConfig() guard in Tauri file-ops to normalize legacy saves
- Fix collectCell() immutability violation in CLI paint command
- Add JSDoc on engine.getState() documenting empty palette behavior
- Update GitNexus index stats (284 symbols, 804 relationships)
