+++ CLAUDE.md
@@ -83,7 +83,7 @@ Note: The return object currently has duplicate keys (`setCurrentImage` x2, `onB
 <!-- gitnexus:start -->
 # GitNexus — Code Intelligence
 
-This project is indexed by GitNexus as **TrKixel** (245 symbols, 672 relationships, 20 execution flows). Use the GitNexus MCP tools to understand code, assess impact, and navigate safely.
+This project is indexed by GitNexus as **TrKixel** (284 symbols, 804 relationships, 20 execution flows). Use the GitNexus MCP tools to understand code, assess impact, and navigate safely.
 
 > If any GitNexus tool warns the index is stale, run `npx gitnexus analyze` in terminal first.
